Plutarch.Test

Description

Common functions for testing Plutarch code

Plutarch specific Expectation operators

passert :: ClosedTerm a -> Expectation

Asserts the term to be true

passertNot :: ClosedTerm a -> Expectation

Asserts the term to be false

pfails :: ClosedTerm a -> Expectation

Asserts the term evaluates without success

psucceeds :: ClosedTerm a -> Expectation

Asserts the term evaluates successfully without failing

ptraces :: ClosedTerm a -> [Text] -> Expectation

Asserts that the term evaluates successfully with the given trace sequence

pshouldBe :: ClosedTerm a -> ClosedTerm b -> Expectation

Like shouldBe but but for Plutarch terms

(#@?=) :: ClosedTerm a -> ClosedTerm b -> Expectation

Like @?= but for Plutarch terms

Budget expectation

psatisfyWithinBenchmark :: Benchmark -> Benchmark -> Expectation

Check that the given benchmark is within certain maximum values.

Use this to ensure that a program's benchmark doesn't exceed expected values (such as script size or memory budget). You will need this because,

Test's golden testing uses maximum possible ExBudget for evaluating programs

You may want to check that the script size is within a certain value

Golden testing

(@|) :: forall t a. HasGoldenValue t => GoldenKey -> (forall s. t s a) -> PlutarchGoldens infixr 0

Specify goldens for the given Plutarch program

(@\) :: GoldenKey -> PlutarchGoldens -> PlutarchGoldens

Add an expectation for the Plutarch program specified with (@|)

(@->) :: ClosedTerm a -> (ClosedTerm a -> Expectation) -> TermExpectation a infixr 1

Test an expectation on a golden Plutarch program

(@:->) :: ClosedTerm a -> ((ClosedTerm a, Script, Benchmark) -> Expectation) -> TermExpectation a infixr 1

Like @-> but also takes the evaluated script and benchmark as arguments

Useful to do assertion checks on post-evaluation benchmark (eg: to check if script size is below certain threshold) -- use in conjunction with psatisfyWithinBenchmark -- or on evaluated script (ie., without re-evaluating the program).

(@==) :: ClosedTerm a -> ClosedTerm b -> TermExpectation a infixr 1

Test that the Plutarch program evaluates to the given term

pgoldenSpec :: HasCallStack => PlutarchGoldens -> Spec

Create golden specs for pre/post-eval UPLC and benchmarks.

A *single* golden file will be created (for each metric) for all the programs in the given tree.

For example, ``` pgoldenSpec $ do "foo" | pconstant 42 "bar" do "qux" @| pconstant Hello ```

Will create three golden files -- uplc.golden, uplc.eval.golden and bench.golden -- each containing three lines one for each program above. Hierarchy is represented by intercalating with a dot; for instance, the key for qux will be "bar.qux".

pgoldenSpec' :: HasCallStack => GoldenConf -> PlutarchGoldens -> Spec

Like pgoldenSpec but takes a GoldenConf to determine which goldens to track.

pgoldenSpec = pgoldenSpec' def

type PlutarchGoldens = ListSyntax (GoldenKey, GoldenValue)

data GoldenConf

Constructors

GoldenConf

Fields

chosenTests :: Set GoldenTest

goldenBasePath :: FilePath

Directory to put the goldens in.

data GoldenTest

Constructors

GoldenT'UPLCPreEval
The unevaluated UPLC (compiled target of Plutarch term)

GoldenT'UPLCPostEval
The evaluated UPLC (evaluated result of Plutarch term)

GoldenT'Bench
Benchmark of Plutarch term (will never fail)

Benchmark type for use in (@:->)

data Benchmark

Constructors

Benchmark

Fields

exBudgetCPU :: ExCPU

CPU budget used by the script.

exBudgetMemory :: ExMemory

Memory budget used by the script.

scriptSizeBytes :: ScriptSizeBytes

Size of Plutus script in bytes

data ScriptSizeBytes

Test runner related utilities

noUnusedGoldens :: [SpecTree ()] -> IO ()

Ensures that there are no unused goldens left behind.

Use this on any SpecTree that internally uses pgoldenSpec to define the golden tests. These golden file paths are accumulated, and compared to the actual files existing on disk. If any golden file exists on disk, but is not tracked by the SpecTree this function will fail, reporting the list of untracked golden files.

Example:

 noUnusedGoldens =<< hspecAndReturnForest spec
 

noUnusedGoldens' :: GoldenConf -> [SpecTree ()] -> IO ()

Like noUnusedGoldens but takes a custom path to the golden storage.

NOTE: This relies on the assumption that the same GoldenConf is used in all pgoldenSpec' calls. This function will go away after https:/github.comPlutonomiconplutarchissues/458

hspecAndReturnForest :: Spec -> IO [SpecTree ()]

Like hspec, but returns the test forest after running the tests.

Based on https://github.com/hspec/hspec/issues/649#issuecomment-1092423220