Plutarch.Test.Golden

DSL

pgoldenSpec :: HasCallStack => PlutarchGoldens -> Spec

Create golden specs for pre/post-eval UPLC and benchmarks.

A *single* golden file will be created (for each metric) for all the programs in the given tree.

For example, ``` pgoldenSpec $ do "foo" | pconstant 42 "bar" do "qux" @| pconstant Hello ```

Will create three golden files -- uplc.golden, uplc.eval.golden and bench.golden -- each containing three lines one for each program above. Hierarchy is represented by intercalating with a dot; for instance, the key for qux will be "bar.qux".

pgoldenSpec' :: HasCallStack => GoldenConf -> PlutarchGoldens -> Spec

Like pgoldenSpec but takes a GoldenConf to determine which goldens to track.

pgoldenSpec = pgoldenSpec' def

(@|) :: forall t a. HasGoldenValue t => GoldenKey -> (forall s. t s a) -> PlutarchGoldens infixr 0

Specify goldens for the given Plutarch program

(@\) :: GoldenKey -> PlutarchGoldens -> PlutarchGoldens

Add an expectation for the Plutarch program specified with (@|)

(@->) :: ClosedTerm a -> (ClosedTerm a -> Expectation) -> TermExpectation a infixr 1

Test an expectation on a golden Plutarch program

(@:->) :: ClosedTerm a -> ((ClosedTerm a, Script, Benchmark) -> Expectation) -> TermExpectation a infixr 1

Like @-> but also takes the evaluated script and benchmark as arguments

Useful to do assertion checks on post-evaluation benchmark (eg: to check if script size is below certain threshold) -- use in conjunction with psatisfyWithinBenchmark -- or on evaluated script (ie., without re-evaluating the program).

type TermExpectation a = forall s. TermExpectation' s a

type PlutarchGoldens = ListSyntax (GoldenKey, GoldenValue)

Golden key and path

data GoldenKey

The key used in the .golden files containing multiple golden values

currentGoldenKey :: HasCallStack => SpecM () GoldenKey

goldenKeyString :: GoldenKey -> String

mkGoldenKeyFromSpecPath :: HasCallStack => [Text] -> GoldenKey

defaultGoldenBasePath :: FilePath

goldenTestPath :: FilePath -> GoldenKey -> GoldenTest -> FilePath

Golden config

data GoldenConf

Constructors

GoldenConf

Fields

chosenTests :: Set GoldenTest

goldenBasePath :: FilePath

Directory to put the goldens in.

data GoldenTest

Constructors

GoldenT'UPLCPreEval
The unevaluated UPLC (compiled target of Plutarch term)

GoldenT'UPLCPostEval
The evaluated UPLC (evaluated result of Plutarch term)

GoldenT'Bench
Benchmark of Plutarch term (will never fail)

Evaluation

evalScriptAlwaysWithBenchmark :: Script -> (Script, Benchmark)

Like evalScript but doesn't throw EvalError, and returns Benchmark.

On EvalError, this function returns perror as evaluated script. Plutus does not provide an accurate way to tell if the program evaluates to Error or not; see https://github.com/input-output-hk/plutus/issues/4270

compileD :: ClosedTerm a -> Script