Plutarch.Test.Property.HaskEquiv

Description

The property of Plutarch terms corresponding to a Haskell term.

Assuming Haskell functions are already well-tested, by verifying the property that a Plutarch term functions equivalently to the corresponding Haskell term we automatically (more or less) verify the correctness of the Plutarch term.

This modules provides a prop_haskEquiv to that end.

The principal property of the module

prop_haskEquiv :: forall (e :: Equality) (t :: Totality) h p. HaskEquiv e t h p (LamArgs h) => h -> ClosedTerm p -> NP Gen (LamArgs h) -> Property

The given Plutarch term is equivalent to the given Haskell type upto the given Equality and Totality.

Generator arguments must be non-empty if the term is a lambda. This function must always be called using TypeApplications specifying the first two type variables.

Example:

>>> prop_haskEquiv
  @'OnPEq
  @'TotalFun
  (reverse :: [Integer] -> [Integer])
  preverse
  (genList genInteger :* Nil)

data Equality

The nature of equality between two Plutarch terms.

Constructors

OnPEq

OnPData

OnBoth
Terms are equal on both PEq and PData

data Totality

Whether a function is total or partial.

Constructors

TotalFun

PartialFun
The Plutarch *and* Haskell function is expected to be partial (error's on certain inputs).

data NP (a :: k -> Type) (b :: [k]) where

Constructors

Nil :: forall {k} (a :: k -> Type). NP a ('[] :: [k])

(:*) :: forall {k} (a :: k -> Type) (x :: k) (xs :: [k]). a x -> NP a xs -> NP a (x ': xs)

For writing helper functions using prop_haskEquiv

type family LamArgs f :: [Type] where ...

Argument types for a Haskell function (empty if a term value)

Equations

LamArgs (a -> b) = a ': LamArgs b

LamArgs _ = '[]

class LamArgs h ~ args => HaskEquiv (e :: Equality) (t :: Totality) h p args | h -> p

Class of pairs of Plutarch and Haskell types that are semantically equivalent, upto the given Equality and Totality.

Minimal complete definition

haskEquiv

Underlying equality tests

testDataEq :: PIsData a => ClosedTerm a -> ClosedTerm a -> PropertyT IO ()

testPEq :: PEq a => ClosedTerm a -> ClosedTerm a -> PropertyT IO ()