Plutarch.Test.Run

Documentation

noUnusedGoldens :: [SpecTree ()] -> IO ()

Ensures that there are no unused goldens left behind.

Use this on any SpecTree that internally uses pgoldenSpec to define the golden tests. These golden file paths are accumulated, and compared to the actual files existing on disk. If any golden file exists on disk, but is not tracked by the SpecTree this function will fail, reporting the list of untracked golden files.

Example:

 noUnusedGoldens =<< hspecAndReturnForest spec
 

noUnusedGoldens' :: GoldenConf -> [SpecTree ()] -> IO ()

Like noUnusedGoldens but takes a custom path to the golden storage.

NOTE: This relies on the assumption that the same GoldenConf is used in all pgoldenSpec' calls. This function will go away after https:/github.comPlutonomiconplutarchissues/458

hspecAndReturnForest :: Spec -> IO [SpecTree ()]

Like hspec, but returns the test forest after running the tests.

Based on https://github.com/hspec/hspec/issues/649#issuecomment-1092423220